Update off-season animated spinner logo
After a quick chat with lemon, we have agreed that the spinner logo doesn't really work due to a mismatch between the official logo's colour scheme and the pydis background blue. The logo was therefore adjusted to only spin, the colour change is removed. There are also some minor adjustments to the spinning animations that make it look more lively. This also removes some of the weird artefacting we were seeing with the previous version.