Add attribution, and a compatible license for lemojis.
We need to attribute Twemoji since we're remixing and adapting their
emojis to create our Lemojis. These also need to be licensed under a CC BY 4.0
license, which isn't compatible with the one for the rest of our stuff.

This commit takes care of all that, and cleans up the README while at it.