Events: populate with 'meta.md' files
For now, the files contain the start and end dates, and a header.
Descriptions will be added later.

The Evergreen season is a special case, registering as a 'fallback',
rather than having a defined period. This is the branding that we
use when no other event is active.

By default, we understand the 'fallback' key to be False.