Archive hacktoberfest branding
We have decided not to participate in hacktoberfest this year. As such, the
assets created for this events have been archived, to a folder that will
not be picked up by the branding manager in @Python.

An issue will be created for generic Halloween branding to be created and
formalised to work with the branding manager.