Update Halloween & Diwali for 2024
Halloween is on the 31st of October, Diwali is October 30 - November 03.
To deal with the clash, Halloween has been set to end on the 31st of October, and Diwali is set to start on the 1st.
Halloween is given an extra day before, and Diwali an extra day after to best account for timezones that overlap allows.